move table header with background to table tabs
